Fix an issue where nav was not collapsing on route change.
